rework backend texture upload APIs
The main goal is to allow more flexibility, allow cubemap arrays in 
the future and better match vk and metal apis.

Main changes:
- remove updateCubeImage
- remove update2DImage
- update3DImage is now the only texture upload backend API

cubemaps are now treated just like a 2D array of 6 layers.

For this reason, Texture::setImage(..., FaceOFfsets) is deprecated.
Additionally, the 2D versions of Texture::setImage() become inline
helpers.


A side effect of this change is that it is now possible to update only
a single face of a cubemap, but also a region of a face (or faces).